moved helper into query dir
