+++ packages/smooth_app/lib/pages/preferences/user_preferences_dev_mode.dart
@@ -8,7 +8,6 @@ import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/database/dao_product_list.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/database/products_preload_helper.dart';
 import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
 import 'package:smooth_app/generic_lib/loading_dialog.dart';
 import 'package:smooth_app/helpers/data_importer/product_list_import_export.dart';
@@ -19,6 +18,7 @@ import 'package:smooth_app/pages/preferences/user_preferences_dialog_editor.dart
 import 'package:smooth_app/pages/preferences/user_preferences_page.dart';
 import 'package:smooth_app/pages/scan/ml_kit_scan_page.dart';
 import 'package:smooth_app/query/product_query.dart';
+import 'package:smooth_app/query/products_preload_helper.dart';
 
 /// Collapsed/expanded display of "dev mode" for the preferences page.
 ///
